componente de grafico de barra
